feat: design the home section
create the structure for the portrait, the name and job title, the social
and scroll icons.
design a rounded element for the portrait image and add a transparent effect
to it.
add social icons, design them and optimize svg's. To to that add svgo
package into the Docker image.
add an optimized avatar png image.
add the scroll icon and animate it.