feat: create a fake console module
create a little module to be able to create a fake terminal emulator.
this fake console will take a div element and will write lines as a
system terminal.
each character will appear on the screen like a machine write them on
it.
add function to be able to create lines and push each character one by
one inside them.
create a function to display the 404 message inside the 404.html page.
add jest to write unit tests.