feat: create smoothscroll module
create a module to be able to smooth scroll to an anchor element.
when clicking on a link with an anchor, you can call smoothscroll function,
it will apply a smooth scroll with an animation.
the module can take the click event extract the target, his position and
smooth scroll from the current position to the target position.