feat: create menu interactions
add html stuff and class functions to be able to interact with the menu.
add click interactions for each anchor, and use the smoothscroll module to
create the scroll effect on the corresponding element.
manage the scroll to change the active element in the menu when you scroll
to different elements in the DOM.
create a utils module to add reusable function between multiple modules.
refactor smoothscroll module to use useful utils functions.
manage the page change when an element has to redirect to a page and not
an anchor.
add on click to the headers of each page to return to the home page.