feat: fullscreen panels with scroll
implement a fullscreen panel module with a scrolling system.
add css rules with media query to manage each part of the website as panels,
and add the scroll animation.
load a different js module with a media query to be able to manage the
mobile menu and the fullscreen panel system.

implement a function to create and insert a menu for panels.
this menu creates one entry per section and adds a title with the name of
each section.
when it scrolls it update the active element of this menu.
the user can click on each element to scroll to the corresponding section.