feat: create iconLoader function
create iconLoader utils function to display an existing loader element
to the screen when calling network api.
after calling a network api with a fetcher function, iconLoader will
call a render function to dynamicly create element with api response
content.
when the loader is on error mode, stop the animation and add an onclick
if it is defined.