feat: design snackbar
design snackbar to display a short message for the user.
add rules for desktop.
add some animation to display and hide it.
add js code to display the snackbar.
add the snackbar inside the contact section (for a better display on
desktop) because we will use it only on contact section for now.